refactor: move interfaces from the models folder to the interface folder within common
