+++ src/application/authentication/auth.module.ts
@@ -5,7 +5,6 @@ import { PassportModule } from '@nestjs/passport';
 import { AppConfig } from '../../common/interfaces/app-config';
 import { LoginValidationMiddleware } from '../../common/middlewares/login-validation.middleware';
 import { CandidateUserPostgresRepository } from '../../common/repositories/candidate-user/candidate-user-postgres.repository';
-import { CandidateUserSerialize } from '../../common/serializers/candidate-user.serialize';
 import { MailService } from '../../mails/mail.service';
 import { CandidateUserModule } from '../candidate-user/candidate-user.module';
 import { CandidateUserService } from '../candidate-user/candidate-user.service';
@@ -16,6 +15,7 @@ import { GoogleStrategy } from './strategies/google.strategy';
 import { JwtStrategy } from './strategies/jwt.strategy';
 import { LocalStrategy } from './strategies/local.strategy.ts';
 import { CandidateUserSqliteRepository } from '../../common/repositories/candidate-user/candidate-user-sqlite.repository';
+import { CandidateUserSerialize } from '../../common/serializers/candidate-user.serialize';
 
 @Module({
   imports: [
