+++ src/common/decorators/has-roles.decorator.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { Role } from '../../models/roles.enum';
+import { Role } from '../interfaces/authentication/roles.enum';
 
 export const HasRoles = (...roles: Role[]) => SetMetadata('roles', roles);
