+++ src/common/interfaces/authentication/auth-provider.enum.ts
