+++ src/common/interfaces/authentication/auth-request.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
-import { CandidateUserInterface } from '../common/interfaces/candidate-user.interface';
+import { CandidateUserInterface } from '../candidate-user/candidate-user.interface';
 
 export interface AuthRequest extends Request {
   user: CandidateUserInterface;
