+++ src/common/interfaces/candidate-projects/candidate-project.ts
