+++ src/common/interfaces/candidate-user/candidate-token.ts
