+++ src/common/interfaces/candidate-user/candidate-user-from-Jwt.ts
