+++ src/common/interfaces/candidate-user/candidate-user-payload.ts
