+++ src/common/interfaces/candidate-user/candidate-user.interface.ts
