+++ src/common/middlewares/login-validation.middleware.ts
@@ -5,7 +5,7 @@ import {
 } from '@nestjs/common';
 import { validate } from 'class-validator';
 import { NextFunction, Request, Response } from 'express';
-import { LoginRequestBody } from '../../models/login-request-body';
+import { LoginRequestBody } from '../interfaces/authentication/login-request-body';
 
 @Injectable()
 export class LoginValidationMiddleware implements NestMiddleware {
