+++ src/common/repositories/candidate-user/candidate-user-postgres.repository.ts
@@ -5,17 +5,17 @@ import { UpdateCandidateUserDto } from '../../../application/candidate-user/dto/
 import { CandidateUser } from '../../../application/candidate-user/entities/candidate-user.entity';
 import { CandidateUserRepository } from '../../../application/candidate-user/repositories/candidate-user.repository';
 import { PrismaPostgresService } from '../../../config/database/prisma/prisma-postgres.service';
-import { AuthProviderType } from '../../../models/auth-provider.enum';
-import { Role } from '../../../models/roles.enum';
+import { AuthProviderType } from '../../interfaces/authentication/auth-provider.enum';
 import { SALT_BCRYPT } from '../../constants/constants';
 import { CandidateUserSerialize } from '../../serializers/candidate-user.serialize';
+import { Role } from '../../interfaces/authentication/roles.enum';
 @Injectable()
 export class CandidateUserPostgresRepository
   implements CandidateUserRepository
 {
   constructor(
-    private readonly prisma: PrismaPostgresService,
     private readonly candidateUserSerialize: CandidateUserSerialize,
+    private readonly prisma: PrismaPostgresService,
   ) {}
 
   async create(candidate: CreateCandidateUserDto): Promise<CandidateUser> {
