+++ src/common/serializers/candidate-user.serialize.ts
@@ -1,4 +1,4 @@
-import { CandidateUserInterface } from '../interfaces/candidate-user.interface';
+import { CandidateUserInterface } from '../interfaces/candidate-user/candidate-user.interface';
 
 export class CandidateUserSerialize {
   requestToDb(candidate: CandidateUserInterface) {
